AI-Engine/requirements.txt
google-generativeai
faiss-cpu
numpy
python-dotenv